refactor: move CasbinRule into modules
