+++ casbin_pymongo_adapter/adapter.py
@@ -1,41 +1,7 @@
 from casbin import persist
 from pymongo import MongoClient
 
-
-class CasbinRule:
-    """
-    CasbinRule model
-    """
-
-    def __init__(
-        self, ptype=None, v0=None, v1=None, v2=None, v3=None, v4=None, v5=None
-    ):
-        self.ptype = ptype
-        self.v0 = v0
-        self.v1 = v1
-        self.v2 = v2
-        self.v3 = v3
-        self.v4 = v4
-        self.v5 = v5
-
-    def dict(self):
-        d = {"ptype": self.ptype}
-
-        for value in dir(self):
-            if (
-                getattr(self, value) is not None
-                and value.startswith("v")
-                and value[1:].isnumeric()
-            ):
-                d[value] = getattr(self, value)
-
-        return d
-
-    def __str__(self):
-        return ", ".join(self.dict().values())
-
-    def __repr__(self):
-        return '<CasbinRule :"{}">'.format(str(self))
+from ._rule import CasbinRule
 
 
 class Adapter(persist.Adapter):
